+++ src/aptl/backends/aces_realization_model.py
@@ -1,0 +1,115 @@
+"""Structured APTL realization data produced from ACES resources."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+from aces_contracts.diagnostics import Diagnostic
+
+
+@dataclass(frozen=True)
+class NodeRealization:
+    """APTL realization data for one ACES node resource."""
+
+    address: str
+    name: str
+    aliases: tuple[str, ...]
+    profiles: tuple[str, ...]
+    services: tuple[str, ...]
+    rendered_configs: tuple[str, ...]
+    evidence_paths: tuple[str, ...]
+    telemetry_paths: tuple[str, ...]
+    networks: tuple[str, ...]
+    static_addresses: tuple[str, ...]
+
+    def details(self) -> dict[str, object]:
+        return {
+            "address": self.address,
+            "name": self.name,
+            "aliases": list(self.aliases),
+            "profiles": list(self.profiles),
+            "services": list(self.services),
+            "rendered_configs": list(self.rendered_configs),
+            "evidence_paths": list(self.evidence_paths),
+            "telemetry_paths": list(self.telemetry_paths),
+            "networks": list(self.networks),
+            "static_addresses": list(self.static_addresses),
+        }
+
+
+@dataclass(frozen=True)
+class NetworkRealization:
+    """APTL realization data for one ACES network resource."""
+
+    address: str
+    name: str
+    cidr: str | None
+    gateway: str | None
+    internal: bool | None
+
+    def details(self) -> dict[str, object]:
+        return {
+            "address": self.address,
+            "name": self.name,
+            "cidr": self.cidr,
+            "gateway": self.gateway,
+            "internal": self.internal,
+        }
+
+
+@dataclass(frozen=True)
+class PlacementRealization:
+    """APTL realization data for one node-scoped provisioning binding."""
+
+    address: str
+    resource_type: str
+    name: str
+    target_address: str
+    target_node: str | None
+
+    def details(self) -> dict[str, object]:
+        return {
+            "address": self.address,
+            "resource_type": self.resource_type,
+            "name": self.name,
+            "target_address": self.target_address,
+            "target_node": self.target_node,
+        }
+
+
+@dataclass(frozen=True)
+class AptlRealization:
+    """Result of interpreting ACES provisioning content for APTL."""
+
+    profiles: frozenset[str]
+    nodes: tuple[NodeRealization, ...]
+    networks: tuple[NetworkRealization, ...]
+    placements: tuple[PlacementRealization, ...]
+    diagnostics: tuple[Diagnostic, ...]
+
+    def details(self) -> dict[str, object]:
+        resource_counts = {
+            "account-placement": sum(
+                placement.resource_type == "account-placement"
+                for placement in self.placements
+            ),
+            "content-placement": sum(
+                placement.resource_type == "content-placement"
+                for placement in self.placements
+            ),
+            "feature-binding": sum(
+                placement.resource_type == "feature-binding"
+                for placement in self.placements
+            ),
+            "network": len(self.networks),
+            "node": len(self.nodes),
+        }
+        return {
+            "profiles": sorted(self.profiles),
+            "resource_counts": {
+                key: value for key, value in sorted(resource_counts.items()) if value
+            },
+            "nodes": [node.details() for node in self.nodes],
+            "networks": [network.details() for network in self.networks],
+            "placements": [placement.details() for placement in self.placements],
+        }
